Add conversation state tracking. The state of conversation is modeled as a finite state machine, with only 2 states implemented for example purposes:     a) NEW_USER     b) INITIAL with the transitions possible in both directions.
Depending on the current state, the bot's answers will differ.

The conversation state is a property of a user DB model.
This has a limitation for using the bot in group chats,
but so far let's keep it simple - the bot is most likely be
used only in private 1-to-1 chats, so it's not a problem.